requirements.txt
nltk
PyYAML
pytest
